+++ vega/errors.go
@@ -19,8 +19,6 @@ var (
 	ErrPeggedOrderMustBeGTTOrGTC                   = OrderError_ORDER_ERROR_MUST_BE_GTT_OR_GTC
 	ErrPeggedOrderWithoutReferencePrice            = OrderError_ORDER_ERROR_WITHOUT_REFERENCE_PRICE
 	ErrPeggedOrderBuyCannotReferenceBestAskPrice   = OrderError_ORDER_ERROR_BUY_CANNOT_REFERENCE_BEST_ASK_PRICE
-	ErrPeggedOrderOffsetMustBeLessOrEqualToZero    = OrderError_ORDER_ERROR_OFFSET_MUST_BE_LESS_OR_EQUAL_TO_ZERO
-	ErrPeggedOrderOffsetMustBeLessThanZero         = OrderError_ORDER_ERROR_OFFSET_MUST_BE_LESS_THAN_ZERO
 	ErrPeggedOrderOffsetMustBeGreaterOrEqualToZero = OrderError_ORDER_ERROR_OFFSET_MUST_BE_GREATER_OR_EQUAL_TO_ZERO
 	ErrPeggedOrderSellCannotReferenceBestBidPrice  = OrderError_ORDER_ERROR_SELL_CANNOT_REFERENCE_BEST_BID_PRICE
 	ErrPeggedOrderOffsetMustBeGreaterThanZero      = OrderError_ORDER_ERROR_OFFSET_MUST_BE_GREATER_THAN_ZERO
@@ -108,10 +106,6 @@ func (err OrderError) Error() string {
 		return "OrderError: invalid time in force"
 	case OrderError_ORDER_ERROR_BUY_CANNOT_REFERENCE_BEST_ASK_PRICE:
 		return "OrderError: buy cannot reference best ask price"
-	case OrderError_ORDER_ERROR_OFFSET_MUST_BE_LESS_OR_EQUAL_TO_ZERO:
-		return "OrderError: offset must be <= 0"
-	case OrderError_ORDER_ERROR_OFFSET_MUST_BE_LESS_THAN_ZERO:
-		return "OrderError: offset must be < 0"
 	case OrderError_ORDER_ERROR_OFFSET_MUST_BE_GREATER_OR_EQUAL_TO_ZERO:
 		return "OrderError: offset must be >= 0"
 	case OrderError_ORDER_ERROR_SELL_CANNOT_REFERENCE_BEST_BID_PRICE: