+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [212](https://github.com/vegaprotocol/protos/pull/212) - Add pagination field to DelegationRequest
 - [213](https://github.com/vegaprotocol/protos/pull/217) - Add changes for ValidatorPerformance snapshot and events
 - [219](https://github.com/vegaprotocol/protos/pull/219) - Include floating point consensus flags in the snapshot for a market
+- [189](https://github.com/vegaprotocol/protos/pull/189) - Update offset to string and disable negative offset for buy side
 
 ### 🐛 Fixes
 - [202](https://github.com/vegaprotocol/protos/pull/202) - Replaces Withdrawal Status Cancelled with Rejected which is more accurate.